Alterando de int para long

+++ src/main/java/br/task/task/model/Task.java
@@ -1,5 +1,7 @@
 package br.task.task.model;
 
+import java.util.function.LongFunction;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,7 +20,7 @@ public class Task {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String title; 
     private String description;
     private Boolean completed = false;
